requirements.txt
# Core Agent Libraries
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.3.3
ollama>=0.1.7

# Vector Store
qdrant-client>=1.7.0

# Web Integration Tools
serpapi==0.1.5
requests>=2.31.0

# Essentials
python-dotenv>=1.0.0
numpy>=1.24.3
